requirements.txt
neo4j==5.24.0
pandas==2.2.2
pyyaml==6.0.2
numpy==1.26.4
pytest==8.3.2
matplotlib==3.9.2
scikit-learn==1.3.2
pyarrow-23.0.0